Some state machine work
Trying to implement state machine with the player. Since all of his code is in 1 place, and it's not sustainable long-term, when adding mechanics. Also, I'm adding an Actions class, since different actions are starting to pile up, and I wanna keep the main class as clean as possible, and keep all my actions in 1 dedicated place.