Dash effect added
So I've done quite a few things with this, and I probably shouldn't have done it all in some random branch, but it is what it is. Dash effect is done, kinda. I also added controller support, using Unity's input system, which took me a while to figure out, since most uses for it are via events, which I don't want.